% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corine_map.R
\name{corine_map}
\alias{corine_map}
\title{Show Corine roughness on county on the map.}
\usage{
corine_map(code_je = "0414")
}
\arguments{
\item{code_je}{character vector; code of Polish county}
}
\value{
show a leaflet map with Corine roughness areas.
}
\description{
Display Corine roughness areas of a selected county on the map.
}
\examples{
corine_map()

}
