% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/county_map.R
\name{county_map}
\alias{county_map}
\title{Show county on the map.}
\usage{
county_map(code_je = "1465")
}
\arguments{
\item{code_je}{character vector; code of Polish county}

\item{county}{character vector; name of Polish county}
}
\value{
show a leaflet map with a county border.
}
\description{
Display borders of a selected county on the map.
}
\examples{
county_map()

}
