% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_wasp_map.R
\name{export_wasp_map}
\alias{export_wasp_map}
\title{WAsP map based on CORINE map.}
\usage{
export_wasp_map(fname = "", code_je = "0414", wgs_zone = 0, buffer = -1)
}
\arguments{
\item{fname}{character vector; file name}

\item{code_je}{character vector; code of Polish county}

\item{buffer}{numeric vector; negative shape buffer value}

\item{wsg_zone}{numeric vector; 33 or 34 UTM WGS84 zone}
}
\value{
show a leaflet map with Corine roughness areas.
}
\description{
Extract and export to a file roughness areas WAsP map based on Corine map.
}
\examples{
export_wasp_map()

}
