% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_clc.R
\name{load_clc}
\alias{load_clc}
\title{Load Corine land cover definition data frame.}
\usage{
load_clc(fname = "")
}
\arguments{
\item{fname}{character vector; file name}
}
\description{
Load Corine land cover definition data frame from csv file.
}
\examples{
load_clc()

}
