% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_clc.R
\name{save_clc}
\alias{save_clc}
\title{Save Corine land cover definition data frame.}
\usage{
save_clc(fname = "")
}
\arguments{
\item{fname}{character vector; file name}
}
\description{
Save Corine land cover definition data frame to csv file.
}
\examples{
save_clc()

}
