% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_clc.R
\name{show_clc}
\alias{show_clc}
\title{Show CORINE Land Cover leveles.}
\usage{
show_clc()
}
\value{
a tibble CORINE Land Cover leveles.
}
\description{
Display a data frame with CORINE Land Cover leveles, roughness length vaulesa
and descriptions in Polish and English.
}
\examples{
show_teryt()

}
