% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_teryt.R
\name{show_teryt}
\alias{show_teryt}
\title{Show TERYT data.}
\usage{
show_teryt(teryt = "", voivodeship = "", code_je = "", county = "")
}
\arguments{
\item{teryt}{character vector; two digits TERYT number, code of Polish voivodeship}

\item{voivodeship}{character vector; Polish voivodeship name}

\item{code_je}{charatcer vector; code of Polish county}

\item{county}{character vector; name of Polish county}
}
\value{
a tibble with requested information
}
\description{
Display TERYT code, voivodship, code_je of county and county name.
}
\examples{
show_teryt()

}
